Update Controllers: DirectorController, ActorController

+++ app/Http/Controllers/ActorController.php
@@ -3,51 +3,69 @@
 namespace App\Http\Controllers;
 
 use App\Http\Controllers\Controller;
+use App\Models\Actors;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Validator;
 
 class ActorController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth:api', ['except' => ['index','store','update', 'delete']]);
+    }
+
     public function index()
     {
-        $actors = Actors::all();
-        return view('actors.index', compact('actors'));
+        $actors = Actors::get();
+        return response()->json($actors, 200);
     }
-    public function edit($id)
-    {
-        $actor = Actors::findOrFail($id);
 
-        return view('actors.edit', compact('actor'));
-    }
-    public function destroy($id)
-    {
-        $actor = Actors::findOrFail($id);
-        $actor->delete();
-
-        return redirect()->route('actors.index')
-            ->with('success','Actor deleted successfully');
-    }
     public function store(Request $request)
     {
-        $request->validate([
-            'name' => 'required',
-            'image' => 'required|image|mimes:jpeg,png,jpg,gif|max:2048',
-            'placeOfBirth' => 'required',
-            'birthday' => 'required|date',
-            'description' => 'required',
+        $validator = Validator::make($request->all(), [
+            'name' => 'required|string',
+            'image' => 'nullable|string',
+            'placeOfBirth' => 'nullable|string',
+            'birthday' => 'nullable|string',
+            'description' => 'nullable|string'
         ]);
-
-        $imageName = time().'.'.$request->image->extension();
-        $request->image->move(public_path('images'), $imageName);
-
-        Actors::create([
-            'name' => $request->name,
-            'image' => $imageName,
-            'placeOfBirth' => $request->placeOfBirth,
-            'birthday' => $request->birthday,
-            'description' => $request->description,
+        if($validator->fails()){
+            return response()->json($validator->errors()->toJson(), 400);
+        }
+        $actor = Actors::create([
+            'name' => $request['name'],
+            'image' => $request['image'],
+            'placeOfBirth' => $request['placeOfBirth'],
+            'birthday' => $request['birthday'],
+            'description' => $request['description']
         ]);
+        return response()->json($actor);
+    }
 
-        return redirect()->route('actors.index')
-            ->with('success','Actor created successfully.');
+    public function update(Request $request)
+    {
+        $validatedData = $request->validate([
+            'name' => 'required|string',
+            'image' => 'nullable|string',
+            'placeOfBirth' => 'nullable|string',
+            'birthday' => 'nullable|string',
+            'description' => 'nullable|string'
+        ]);
+        $actor = Actors::where("id",$request["id"])->first();
+        $actor->update($validatedData);
+        return response()->json(['message'=>'Complete!'],200);
+    }
+
+    public function delete(Request $request)
+    {
+        $validator = Validator::make($request->all(), [
+            'id' => 'required|int'
+        ]);
+        if($validator->fails()){
+            return response()->json($validator->errors()->toJson(), 400);
+        }
+        $actor = Actors::where("id",$request["id"])->first();
+        $actor->delete();
+        return response()->json(['message'=>'Complete!'],200);
     }
 }
